-- renamed hexapete -> hexapede -- added new folder for launch files in hexapede
