+++ hexapede/hexapede/package.xml
@@ -1,25 +1,25 @@
 <?xml version="1.0"?>
 <package>
-  <name>hexapete_joy_to_ctrl</name>
+  <name>hexapede</name>
   <version>0.0.0</version>
-  <description>The hexapete_joy_to_ctrl package</description>
+  <description>The hexapede package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag --> 
   <!-- Example:  -->
   <!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
-  <maintainer email="roeckma52465@th-nuernberg.de">roeckma52465</maintainer>
+  <maintainer email="chris@todo.todo">chris</maintainer>
 
 
   <!-- One license tag required, multiple allowed, one license per tag -->
   <!-- Commonly used license strings: -->
   <!--   BSD, MIT, Boost Software License, GPLv2, GPLv3, LGPLv2.1, LGPLv3 -->
-  <license>BSD</license>
+  <license>TODO</license>
 
 
   <!-- Url tags are optional, but mutiple are allowed, one per tag -->
   <!-- Optional attribute type can be: website, bugtracker, or repository -->
   <!-- Example: -->
-  <!-- <url type="website">http://wiki.ros.org/hexapete_joy_to_twist</url> -->
+  <!-- <url type="website">http://wiki.ros.org/hexapede</url> -->
 
 
   <!-- Author tags are optional, mutiple are allowed, one per tag -->
@@ -40,21 +40,11 @@
   <!-- Use test_depend for packages you need only for testing: -->
   <!--   <test_depend>gtest</test_depend> -->
   <buildtool_depend>catkin</buildtool_depend>
-
-  <build_depend>std_msgs</build_depend>
-  <build_depend>geometry_msgs</build_depend>
-  <build_depend>joy</build_depend>
-  <build_depend>roscpp</build_depend>
-  
-  <run_depend>std_msgs</run_depend>
-  <run_depend>geometry_msgs</run_depend>
-  <run_depend>joy</run_depend>
-  <run_depend>roscpp</run_depend>
-
+  <run_depend>hexapete_ctrl_motor</run_depend>
+  <run_depend>hexapete_launch</run_depend>
 
   <!-- The export tag contains other, unspecified, tags -->
   <export>
-    <!-- Other tools can request additional information be placed here -->
-
+    <metapackage/>
   </export>
 </package>
