+++ hexapede/hexapede_launch/package.xml
@@ -1,25 +1,25 @@
 <?xml version="1.0"?>
 <package>
-  <name>hexapete_ctrl_motor</name>
+  <name>hexapede_launch</name>
   <version>0.0.0</version>
-  <description>The hexapete_ctrl_motor package</description>
+  <description>The hexapede_launch package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag --> 
   <!-- Example:  -->
   <!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
-  <maintainer email="roeckma52465@th-nuernberg.de">roeckma52465</maintainer>
+  <maintainer email="chris@todo.todo">chris</maintainer>
 
 
   <!-- One license tag required, multiple allowed, one license per tag -->
   <!-- Commonly used license strings: -->
   <!--   BSD, MIT, Boost Software License, GPLv2, GPLv3, LGPLv2.1, LGPLv3 -->
-  <license>BSD</license>
+  <license>TODO</license>
 
 
   <!-- Url tags are optional, but mutiple are allowed, one per tag -->
   <!-- Optional attribute type can be: website, bugtracker, or repository -->
   <!-- Example: -->
-  <!-- <url type="website">http://wiki.ros.org/studierbot_rescue</url> -->
+  <!-- <url type="website">http://wiki.ros.org/hexapede_launch</url> -->
 
 
   <!-- Author tags are optional, mutiple are allowed, one per tag -->
@@ -40,17 +40,10 @@
   <!-- Use test_depend for packages you need only for testing: -->
   <!--   <test_depend>gtest</test_depend> -->
   <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>joy</build_depend>
-  <build_depend>roscpp</build_depend>
-  <run_depend>joy</run_depend>
-  <run_depend>roscpp</run_depend>
 
 
   <!-- The export tag contains other, unspecified, tags -->
   <export>
-    <!-- You can specify that this package is a metapackage here: -->
-    <!-- <metapackage/> -->
-
     <!-- Other tools can request additional information be placed here -->
 
   </export>
