+++ hexapede/hexapete_ctrl_motor/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package>
-  <name>hexapete_op</name>
+  <name>hexapede_ctrl_motor</name>
   <version>0.0.0</version>
-  <description>The hexapete_op package</description>
+  <description>The hexapete_ctrl_motor package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag --> 
   <!-- Example:  -->
@@ -19,7 +19,7 @@
   <!-- Url tags are optional, but mutiple are allowed, one per tag -->
   <!-- Optional attribute type can be: website, bugtracker, or repository -->
   <!-- Example: -->
-  <!-- <url type="website">http://wiki.ros.org/hexapete_ctrl</url> -->
+  <!-- <url type="website">http://wiki.ros.org/studierbot_rescue</url> -->
 
 
   <!-- Author tags are optional, mutiple are allowed, one per tag -->
@@ -30,7 +30,6 @@
 
   <!-- The *_depend tags are used to specify dependencies -->
   <!-- Dependencies can be catkin packages or system dependencies -->
-
   <!-- Examples: -->
   <!-- Use build_depend for packages you need at compile time: -->
   <!--   <build_depend>message_generation</build_depend> -->
@@ -41,20 +40,17 @@
   <!-- Use test_depend for packages you need only for testing: -->
   <!--   <test_depend>gtest</test_depend> -->
   <buildtool_depend>catkin</buildtool_depend>
-
-  <!--build_depend>message_generation</build_depend-->
+  <build_depend>joy</build_depend>
   <build_depend>roscpp</build_depend>
-  <build_depend>std_msgs</build_depend>
-  <build_depend>geometry_msgs</build_depend>
- 
-  <!--run_depend>message_runtime</run_depend-->
+  <run_depend>joy</run_depend>
   <run_depend>roscpp</run_depend>
-  <run_depend>std_msgs</run_depend>
-  <run_depend>geometry_msgs</run_depend>
 
 
   <!-- The export tag contains other, unspecified, tags -->
   <export>
+    <!-- You can specify that this package is a metapackage here: -->
+    <!-- <metapackage/> -->
+
     <!-- Other tools can request additional information be placed here -->
 
   </export>
