--- hexapete/hexapete_joy_to_ctrl/src/hexapete_joy_to_ctrl_node.cpp
@@ -1,24 +1,0 @@
-#include "ros/ros.h"
-
-#include "joytoctrl.h"
-#include "param.h"
-
-int main(int argc, char** argv){
-  ros::init(argc,argv,NODE_NAME);
-  JoyToCtrl joy_to_ctrl;
-
-  ros::Rate loop(L_RATE);
-
-  while(ros::ok()){
-    joy_to_ctrl.publish();
-
-    ros::spinOnce();
-    loop.sleep();
-  }
-
-  return(0);
-}
-
-
-
-
