--- hexapete/hexapete_joy_to_ctrl/src/param.h
@@ -1,8 +1,0 @@
-#define NODE_NAME "hexapete_joy_to_ctrl_node"
-#define L_RATE 30
-
-#define DEAD_X_P  0.06
-#define DEAD_X_N -0.02 
-#define DEAD_Z_P  0.17 
-#define DEAD_Z_N -0.1
-
